+++ NEWS.md
@@ -1,3 +1,11 @@
+# CDMConnector 2.4.0
+
+- Implement `generateCohortSet` for local cdms (e.g. with omock)
+- Allow catalogs in duckdb connections @lyoganathan
+- Add `cdmFromCohortSet` to create synthetic cdms from Atlas cohort definitions
+- Add `cdmCommentContents` for creating comments with the contents of a CDM in tests and documentation
+- Make `writeSchema` optional in `cdmFromCon` @SulevR
+
 # CDMConnector 2.3.0
 
 - Add empty CDM v5.4 example dataset @ginberg